[FIX] purchase, inter_company_rules: correct picking type
This revision is related to 13df94f8d1218ca1792514fa88eaee2262223098.

With the same environment and use case than the above revision,
the Purchase order picking type was not set using the correct
company, it always used the user current company, instead
of forcing the company to the company for which the PO is created.

Besides, there is an onchange on the picking type
`onchange_picking_type_id`,
which sets values that are important for the continuation
of the flow.

opw-649378