[FIX] delivery: invoice actual shipping cost
Since commit b14a992f512112d39d55062b70148fcec98fa5e3, there was no more
way to invoice the actual price of a shipping (after the transfer).

In order to make that available again, we introduce a new field on the
sale order. This field automatically detects whether a sale order
contains a "delivery" line or not, at sale order confirmation
(action_confirm).

When the delivery order is processed, the actual shipping cost is thus
repercuted on the sale order if and only if the shipping price was not
"forced" at the sale order confirmation.

(and a few refactoring to avoid copy-pasting code lines)