[FIX] web_graph: duplicates header in pivot
For report with duplicated named headers, several columns could get
squashed together.

This fix check if a value is a duplicate or not, and if it is a
duplicates postfix an number allowing to differentiate the two headers
(for the human and the javascript).

This is the exact same solution than #6722
Backport of 75177ab85ebdef4865323528a9804b7f0d40fa54
opw-650000