[FIX] sale, sale_timesheet: multiple fixes
1) Domain of account.analytic.line

When an expense is created with a cost of zero, the quantity delivered
on the SO line is not updated.

2) Delivered quantity updateable if manual tracking

The delivered quantity cannot be manually modified if it is calculated.
Therefore, we need to restrict the condition to a manual tracking.

3) Upselling conditions

Upselling is possible only for a product invoiced on ordered quantities
for which we delivered more than expected.