[IMP] point_of_sale: add remote debugging capabilities to the posbox
Sometimes it's useful to ssh into a posbox to quickly troubleshoot an
issue. Talking customers through port forwarding is very difficult
however. We use ngrok to solve this issue.

This implements an interface that allows a user to enter an ngrok
authentication token. After this, ngrok can be started and we should be
able to remotely connect.

Note that ngrok does NOT run by default on the posbox. It can only be
started manually by the user.