[FIX] website_membership: don't display old as members
Follow-up of 80f4ba0 which solved the issue when an association was
selected, but not when "All Members" was choosen.

Before this commit, the displayed membership lines would only be the
paid ones, if an association (membership product) was selected. With
this commit this is also the case when no such association is selected.

This commit also correct the member counts in the "country" list :

* with "All members" the free members were not in the counts,
* with not "Free member", the counts didn't took the eventual current name search.

closes #9083
opw-648627