[IMP] point_of_sale: stop warning unknown field 'accuracy'
In e9c6fd0 'accuracy' was added to the field list read for res.currency,
but since the accounting refactoring in 9.0 (1a24899), this field was no
more present (replaced by decimal_places).

Since this field in the current point_of_sale javascript code was never
used, this commit removes it from the field list.

sentry-32303830312e3239303433353734