[FIX] ir_qweb: call clean_attachment at the end of save_attachment
.. instead of at the start. The reason is: if any rollback have to occur
during the call to `save_attachments`, we want to avoid rollbacking an
unlink on the ir.attachment table (because we cannot rollback the
removal of a file in the filestore).

This issue is especially annoying when we generate a bundle whose
ir.attachment record has the same hash than the invalidated one (for
instance when one of the source files in the bundle has been touched
but not modified) and of of these rollback occurs: the bundle has been
deleted on the filestore and cannot be served.