[FIX] sale, website_quote: Terms and Conditions
When website_quote is installed, the Terms and Conditions of the quote
template should replace the Terms and Conditions of the company.
However, if the T&C of the quote are not filled, they will be kept
empty. This fix uses the T&C of the company as a backup.

This fix also set the TéC in the lang of the partner.

opw-652390