[FIX] sale, website_sale: send email at sale order confirmation in ecommerce
When confirming a sale order through the ecommerce
(at payment, most of the time),
the confirmation email should be sent at confirmation.

This is the subject of the revision
6980ee060e476dfb46f912b11611d88ba1928654
that was lost with revision
234c2b19990e6593238cb9cb7597b668a240d683

At the moment, the email was sent only for
orders not being automatically confirmed
(transfers, for instance).

opw-653800