[FIX] mail: avoid duplicate entries in model mail.channel.partner
in some cases, it is possible to call the method
channel_join_and_getinfo when the user is already in the channel.  This
may (and did) cause issues in some unrelated code.  This commit
guarantees that this codepath will avoir creating another entry.

Another way to achieve this result is by adding a constraint on the
model mail.channel.partner, but this is in 9.0, so we must be careful.